Add skeleton implementation of single-session API
This adds an SNMPSession class with some expected low-level operations
as methods.  The plan is work with Net-SNMP's single-session API.

For the time being, the implementation is experimental, as part of my
learning process, so to ease into things, this defines a traditional,
synchronous interface, not the asynchronous one we'll be aiming for in
the longer run.