Revert to traditional Session API
It turns out I've misunderstood how to use the single-session API.
Thread-safety doesn't matter to us if we're limited to a single session
per thread.  It's more important to support multiple sessions in a
single thread.  It seems the only way to achieve horizontal scaling
then is to add multiple processes.