Add functions for working with select() calls
This adds utility functions to get information from Net-SNMP on which
file descriptors it is currently responsible for that need to be a part
of a client process' `select()` call (when working in asynchronous
event loops).  This is based on the Net-SNMP's large fd sets.

Large fd sets allow Net-SNMP to work with sets of file descriptors that
are larger than 1024 (which is an oft-imposed limit on the number of
open file descriptions in a single Linux process).  We might as well
use these from the get-go to allow for arbitrarily huge numbers.