Add functions for asyncio event loop updating
The Net-SNMP library will open and keep track of its own socket file
descriptors.  Since we want to support asyncio in our library, we
need to ensure that our event loop knows which Net-SNMP file descriptors
it should add to its watch list (i.e. it must know to react to Net-SNMP
sockets when they have data ready to read).

Net-SNMP itself also keeps track of which SNMP sessions and requests
have timed out.  Through its advisory functions, it will let us know
which file descriptors it needs to add to a `select()` call, but also
which timeouts it is waiting for.

This adds utility functions that keep track of which Net-SNMP file
descriptors are coming and going, and making sure to add and remove
the corresponding reader callbacks to/from the asyncio event loop.