Keep a map of all sessions
We intend to declare a static callback function for Net-SNMP to use
for all responses.  This function needs to be able to map the incoming
response to the proper SNMPSession Python object, so we keep a
class variable for this and ensure we update it and the event loop
every time a session is opened or closed.