Use `asyncio.get_running_loop()`
`asyncio.get_event_loop()` is being deprecated, so we should use
`asyncio.get_running_loop()` instead.

This also adds a guard to `update_event_loop()`, since it may be called
outside of a running event loop on non-asyncio contexts (e.g. opening a
session and then trying to use `asyncio.run()` to use the async
methods).  It doesn't matter if the initial call to
`update_event_loop()` is skipped, as it will be called on every request
that runs in an event loop  as well (it's designed to be run at any
time).