Make package build work
This cleans up the CFFI definitions to be able to build the FFI
interface modules as part of the netsnmp-cffi package build.

Unfortunately, there doesn't seem to be a useful way to tell CFFI
to build our C integration module using pyproject.toml, so this adds
a setup.py to do it.

This also names the low-level module properly as `netsnmpy._netsnmp`,
since it belongs in our package scope.