Support multiple event loops
The `_fd_map` was a global mapping between file descriptors and
`SNMPReader` instances added to the event loop.  However, during
synchronous usage, the event loop may be different for every request,
meaning that subsequent synchronous requests after the first one will
fail: They will run in a new event loop where the SNMPReader has not
been added.

This changes the map into being a map of fd maps *per event loop*.  It
uses a `WeakKeyDictionary` for the outer map, so that when an event loop
ceases to exist, its `SNMPReader` mapping should also cease to exist.

This fixes the regression test added in the parent commit.