Formalize varbinds as the `SNMPVariable` class.
Many values retrieved through SNMP may have a protocol-level data type,
but need to be interpreted in light of the MIB object whose value they
represent.  The new `SNMPVariable` class formalizes the tuple of an
SNMP variable (or varbind): A specific object definition (the OID
representing an object from the MIB) and its retrieved value.

Most importantly, this adds lookups of symbolic names of enumerations
defined in MIBs.  This enables a client to defer the lookup of such
values to the MIB functionality of the Net-SNMP library.