Use bigger buffer for symbol name translation
Some symbols become quite long, and we very quickly ran into symbols
from Juniper's proprietary BGP MIB that were 128 characters or more.

This doubles the size of the C buffer used to receive the symbolic
name, and adds a fallback if the C call indicates that the buffer
was too small to fit the resulting string.