Clean up project requirements specs
The optional dev requirements do not really belong in the section of
optional requirements, as they not optional runtime requirements for the
package: They're just a list of useful or needed tools for working with
the codebase as a developer, and they will never provide any extra
features if installed into runtime.

https://peps.python.org/pep-0735/ has been introduced to alleviate this
issue with a new `[dependency-groups]` section of `pyproject.toml`, but it
is unclear to me how well this is supported in the various build tools
as of yet - which is why this commit relies on "good old" requirements
files.