Fix broken test runs on Python 3.12.
This should ensure the test requirement `snmpsim` runs properly
under Python 3.12.  1.1.6 release had a bug, so we don't want that.
Also, it uses a version of PySNMP that will not work on 3.12 due to
the removal of the stdlib asyncore module, so we need to ensure that
the pyasyncore portability module is present on Python 3.12 for the
tests to work.