Add regression test for non-ascii community traps
A device that sends traps with non-ascii community values will cause
trap PDU processing to crash with a `UnicodeDecodeError`. community
values are defined as ASCII in the SNMP specs, but some devices do not
behave. netsnmp-cffi tries to be helpful to the client by automatically
making a Python string from the community, but if there's a misbehaving
device, this should not break everything.