Cancel pending futures when session is closed
If a session is closed while requests are still pending, Net-SNMP
will never call the session's callbacks again (neither for responses
or timeouts). The only sensible thing to do is to cancel all of the
session's pending futures to ensure all waiting coroutines can resume
with a suitable exception for their trouble.